# docs/openapi.yaml
openapi: 3.1.0
info:
  title: oracle API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: >
    HTTP API for oracle, the self-hosted assistant. Every `/api/v1` route
    requires an authenticated session cookie issued by the auth server
    mounted at `/auth/*` (Limen, out of scope of this document — see the
    auth package and the frontend's `api.ts` for the calls it makes).
    Admin-only routes are marked; non-admins receive 403
    (`{"error":"admin access required"}`).

    Errors are JSON `{"error": "message"}`. The chat and approvals endpoints
    stream Server-Sent Events instead of returning a body.
servers:
  - url: /
    description: Same-origin deployment (the web app proxies /api/v1 to the API).
...
paths:
...
  /api/v1/sessions:
    get:
      tags: [sessions]
      summary: List chat sessions
      description: Only the current user's sessions, newest first.
      operationId: listSessions
      security:
        - sessionCookie: []
      responses:
        '200':
          description: The user's sessions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sessions/{id}:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    patch:
      tags: [sessions]
      summary: Update a session
      description: >
        Rename the session and/or change its goal loop configuration. Only
        fields present in the body are changed. Enabling the loop (or changing
        its interval while enabled) queues the next iteration; disabling
        cancels any queued run.
      operationId: updateSession
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: New session title.
                loop_enabled:
                  type: boolean
                  description: Enable or disable the session's goal loop.
                loop_interval:
                  type: string
                  description: >
                    Delay between loop iterations as a Go duration ("30s",
                    "5m"). Empty string runs continuously (as soon as the
                    previous iteration finishes).
      responses:
        '200':
          description: The updated session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [sessions]
      summary: Delete a session and its messages
      operationId: deleteSession
      security:
        - sessionCookie: []
      responses:
        '204':
          description: Deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sessions/{id}/messages:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    get:
      tags: [sessions]
      summary: List a session's messages
      description: Tool calls are nested per message when present.
      operationId: listMessages
      security:
        - sessionCookie: []
      responses:
        '200':
          description: The session's messages, oldest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie issued by the auth server at /auth/*.
    adminRole:
      type: oauth2
      flows: {}
      description: >
        Marker scheme for admin-only routes. Real enforcement is role-based
        on the session's user (role `admin`); this scheme documents the
        requirement only.
  parameters:
    SessionId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Session id (visible in the start event).
...
  responses:
...
    Unauthorized:
      description: Missing or invalid session.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: The resource does not exist or belongs to another user.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Unexpected failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    SessionInfo:
      type: object
      required: [id, title, created_at, updated_at]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        loop_enabled:
          type: boolean
        loop_interval:
          type: string
          description: Delay between iterations as a Go duration; empty means continuous.
        loop_next_run_at:
          type: [string, 'null']
          format: date-time
          description: When the next iteration is due, null when none is queued.
        loop_last_run_at:
          type: [string, 'null']
          format: date-time
        loop_last_status:
          type: string
          description: running, done, error, or budget (server loop_max_runs cap reached); empty when never run.
        loop_error:
          type: string
          description: Error from the last iteration, empty when none.
        loop_run_count:
          type: integer
          format: int64
          description: Total iterations run on this session.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SessionMessage:
      type: object
      required: [id, role, content, created_at]
      properties:
        id:
          type: integer
          format: int64
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        created_at:
          type: string
          format: date-time
        tool_calls:
          type: array
          items:
            $ref: '#/components/schemas/ToolCallResult'
    ToolCallResult:
      type: object
      required: [id, call_id, name, arguments, result, status]
      properties:
        id:
          type: integer
          format: int64
        call_id:
          type: string
        name:
          type: string
        arguments:
          type: string
          description: JSON-serialized arguments.
        result:
          type: string
        status:
          type: string
          description: success, error, or denied.
...